Kp 11183 updated

+++ kcommons/src/main/java/com/kairos/dto/activity/activity/activity_tabs/TimeCalculationActivityDTO.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.time.DayOfWeek;
 import java.time.LocalTime;
 import java.util.List;
@@ -32,7 +33,7 @@ public class TimeCalculationActivityDTO implements Serializable {
     private Boolean multiplyByVacationFactor;
     private Boolean multiplyByFinalSchedule;
     private String breakTemplates;
-    private List<Long> dayTypes;
+    private List<BigInteger> dayTypes;
     private DayOfWeek fullWeekStart;
     private DayOfWeek fullWeekEnd;
     private int historyDuration;
@@ -174,11 +175,11 @@ public class TimeCalculationActivityDTO implements Serializable {
         this.activityId = activityId;
     }
 
-    public List<Long> getDayTypes() {
+    public List<BigInteger> getDayTypes() {
         return dayTypes;
     }
 
-    public void setDayTypes(List<Long> dayTypes) {
+    public void setDayTypes(List<BigInteger> dayTypes) {
         this.dayTypes = dayTypes;
     }
